requirements.txt
Flask==2.1.0
pytesseract==0.3.10
Pillow==9.0.0
requests==2.28.1
flask-cors==3.1.1